b2FixedJoint. Thanks Jorrit Rouwé.
git-svn-id: https://box2d.svn.sourceforge.net/svnroot/box2d/trunk@220 f71193c7-d439-0410-8131-bb32c6e3d2ad

+++ Source/Collision/Shapes/b2CircleShape.cpp
@@ -111,8 +111,7 @@ float32 b2CircleShape::ComputeSubmergedArea(	const b2Vec2& normal,
 	//Magic
 	float32 r2 = m_radius*m_radius;
 	float32 l2 = l*l;
-    //TODO: write b2Sqrt to handle fixed point case.
-	float32 area = r2 * (asin(l/m_radius) + b2_pi/2.0f)+ l * b2Sqrt(r2 - l2);
+	float32 area = r2 * (asinf(l/m_radius) + b2_pi/2.0f)+ l * b2Sqrt(r2 - l2);
 	float32 com = -2.0f/3.0f*pow(r2-l2,1.5f)/area;
 	
 	c->x = p.x + normal.x * com;
